// src/packages/bui-react/src/components/Image/Image.types.ts
import React from "react";
import { getImageAssetUrl } from "@bookingcom/bui-assets-react";
import type { Responsive, Mixin } from "@bookingcom/bui-core/types";
import type { IconSVG } from "components/Icon";
import type * as G from "types/global";

export type Props = {
  alt?: string;
  src?: string;
  asset?: Parameters<typeof getImageAssetUrl>[0];
  height?: Responsive<number | string>;
  width?: Responsive<number | string>;
  borderRadius?: "100" | "200" | "300";
  contentMode?: "fill" | "fit";
  fallback?: "image" | "icon" | "background" | null;
  fallbackImageSrc?: string;
  fallbackImage?: React.ReactNode;
  fallbackImagePadding?: "none" | "small" | "large";
  fallbackIcon?: IconSVG;
  onError?: () => void;
  priority?: boolean;
  className?: string;
  attributes?: G.Attributes<"picture">;
  imgClassName?: string;
  imgAttributes?: G.Attributes<"img">;
  mixin?: Mixin<{ padding: false; width: false; height: false }>;
};